Clean up and move QueryIndexStatusExecutor to an inner class
Tests for `QueryIndexStatusExecutor` have also been restructured, using directly QueryExecutor.
